testファイルをメソッド毎に分けた
